Update macOS build script to enable runtime, build DMG instead of PKG; update Tor; update build instructions

+++ install/build_osx.sh
@@ -20,9 +20,7 @@ cd $ROOT
 if [ "$1" = "--release" ]; then
   mkdir -p dist
   APP_PATH="$ROOT/dist/OnionShare.app"
-  PKG_PATH="$ROOT/dist/OnionShare.pkg"
-  IDENTITY_NAME_APPLICATION="Developer ID Application: Micah Lee"
-  IDENTITY_NAME_INSTALLER="Developer ID Installer: Micah Lee"
+  IDENTITY_NAME_APPLICATION="Developer ID Application: Micah Lee (N9B95FDWH4)"
   ENTITLEMENTS_CHILD_PATH="$ROOT/install/macos_sandbox/child.plist"
   ENTITLEMENTS_PARENT_PATH="$ROOT/install/macos_sandbox/parent.plist"
 
@@ -30,26 +28,30 @@ if [ "$1" = "--release" ]; then
   codesign \
     --deep \
     -s "$IDENTITY_NAME_APPLICATION" \
+    -o runtime \
     --force \
     --entitlements "$ENTITLEMENTS_CHILD_PATH" \
     --timestamp \
     "$APP_PATH"
   codesign \
     -s "$IDENTITY_NAME_APPLICATION" \
+    -o runtime \
     --force \
     --entitlements "$ENTITLEMENTS_PARENT_PATH" \
     --timestamp \
     "$APP_PATH"
 
-  echo "Creating an installer"
-  productbuild \
-    --sign "$IDENTITY_NAME_INSTALLER" \
-    --component "$APP_PATH" /Applications \
-    --timestamp \
-    "$PKG_PATH"
+  echo "Create the DMG"
+  if [ ! -f "/usr/local/bin/create-dmg" ]; then
+    echo "Error: create-dmg is not installed"
+    exit 0
+  fi
+  /usr/local/bin/create-dmg "$APP_PATH" --identity "$IDENTITY_NAME_APPLICATION"
+  mv *.dmg dist
 
   echo "Cleaning up"
   rm -rf "$APP_PATH"
 
-  echo "All done, your installer is in: $PKG_PATH"
+  echo "All done, your DMG is in:"
+  ls dist/*.dmg
 fi
